requirements.txt
fastapi
uvicorn
pydantic
fastapi
python-multipart
pillow
numpy
opencv-python
insightface
onnxruntime-azure
